Simplify README: less is more
Drop most of the talk and get straight to business.

Instead of describing the code, just expect the code to describe itself.

Move the information about the non-NPM usage at the end of the
README to get it out of the really useful information.

Add a bit longer summary at the top to describe the purpose
of the library in bit more detail.