Drop aliasAs config option
I started to refactor/rewrite our existing implementation,
to only support aliasAs only inside SELECT clause,
but came to the realization that there are so many cases
where this option is broken, and it's not really sensible
to attempt fixing all these issues with the current
token-list-manipulation approach.

Really the better way is to perform this transformation
after parsing.  But for now our AST doesn't contain enough
information for this.