Move knowledge of token types out of TokenizerEngine
TokenizerEngine now takes an array of TokenRules, the ordering
of the array determines the order of tokenization.  The TokenizerEngine
itself no more has any knowledge of which particular tokens its
dealing with, greatly reducing the coupling between Tokenizer
and the -Engine part, as token types no more need to be repeated
in both places.