Allow BETWEEN operator to be followed by AND operator
Fixes a long-standing bug which resulted in ambiguous grammar.

The solution involves adding multiple new parser rules
which exclude the AND/OR/XOR operators from expressions
inside the BETWEEN-expression.

This solution is far from elegant. Really it's plain ugly,
but at least it gets the bug fixed. I think there's no good
way to fix it in the current parser/formatter architecture.

Fixes #534